feat(types): extract logger/lock/deployment types into shared types module
